requirements.txt
django==5.0.7
djangorestframework-api-key==3.0.0
django-cors-headers==4.4.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.2
decorator==4.4.2
distlib==0.3.8
docopt==0.6.2
docutils==0.17
ebaysdk==2.1.5
executing==2.0.1
filelock==3.15.4
freezegun==1.1.0
geoip2==2.9.0
idna==2.10
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.2
jupyter_core==5.7.2
libsass==0.22.0
lxml==5.2.1
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
maxminddb==2.6.2
nest-asyncio==1.6.0
num2words==0.5.10
ofxparse==0.21
packaging==24.1
parso==0.8.4
passlib==1.7.4
pillow==10.2.0
platformdirs==4.2.2
polib==1.1.1
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
pydot==1.4.2
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==2.12.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-stdnum==1.17
pytz==2024.1
pyusb==1.2.1
pyzmq==26.0.3
qrcode==7.3.1
reportlab==4.1.0
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.1.0
rl-renderPM==4.0.3
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
urllib3==2.0.7
virtualenv==20.26.3
vobject==0.9.6.1
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
djangorestframework-simplejwt==5.3.1
django-celery-results==2.5.1
celery==5.4.0
django-celery-beat==2.6.0
kombu==5.3.7
amqp==5.2.0
python-escpos==3.1
weasyprint==62.3
pdf2image==1.17.0
websockets==12.0
imgkit==1.2.3